+++ examples/sdet-jest30/mockstar.test.js
@@ -1,6 +1,6 @@
 // Satisfies: RT-15 (Jest 30 ESM support via --experimental-vm-modules)
 
-import { launch } from '@dhanesh/mockstar';
+import { launch } from '@dhaneshpurohit/mockstar';
 import { fileURLToPath } from 'node:url';
 import { dirname, resolve } from 'node:path';